Time: 2057 ms (5.04%), Space: 16.4 MB (95.74%) - LeetHub

+++ 0611-valid-triangle-number/0611-valid-triangle-number.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int triangleNumber(vector<int>& nums) {
-        long long count = 0;
+        int count = 0;
         int n = nums.size();
         sort(nums.begin(),nums.end());
